Moved fragment and viewmodel to correct spot
