+++ Corona-Warn-App/src/main/java/de/rki/coronawarnapp/datadonation/survey/ui/onboarding/SurveyConsentViewModel.kt
@@ -1,4 +1,4 @@
-package de.rki.coronawarnapp.tracing.ui.details.items.survey
+package de.rki.coronawarnapp.datadonation.survey.ui.onboarding
 
 import androidx.lifecycle.ViewModel
