+++ Corona-Warn-App/src/main/res/layout/survey_consent_fragment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".tracing.ui.details.items.survey.SurveyConsentFragment">
+    tools:context=".datadonation.survey.ui.onboarding.SurveyConsentFragment">
 
     <TextView
         android:layout_width="match_parent"
